+++ Logic/Global.py
@@ -33,7 +33,7 @@ import Paths
 
 import logging
 
-__version__ = '1.1.5'
+__version__ = '1.1.6'
 
 version = __version__